requirements.txt
black==22.1.0
einops==0.4.0
gin-config==0.5.0
gym[atari]==0.19.0
numpy==1.22.2
opencv-python==4.5.5.62
pandas==1.4.1
torch==1.10.2
tqdm==4.62.3
wandb==0.12.10
